requirements.txt
requests
python-dotenv
patool